+++ src/App.tsx
@@ -12,6 +12,15 @@ const App = () => {
                     <Typography variant="h6">Photo Album</Typography>
                 </Toolbar>
             </AppBar>
+            <main>
+                <div>
+                    <Container maxWidth="sm">
+                        <Typography variant="h2" align="center" color="textPrimary" gutterBottom>
+                            Photo Album
+                        </Typography>
+                    </Container>
+                </div>
+            </main>
         </>
     )
 }